+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Improved implementation of `Primitives::ChanceOf`.
+- Improved `EnumerableExtensions::Intersperse` to support randomness.
 
 ## [6.2.1] - 2023-05-26